requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
cmake==4.2.1
contourpy==1.1.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.57.0
idna==3.11
imageio==2.35.1
imgaug==0.4.0
importlib_resources==6.4.5
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
lit==18.1.8
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.23.5
opencv-python==4.8.0.76
packaging==25.0
Pillow==9.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyWavelets==1.4.1
requests==2.32.4
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.7
six==1.17.0
sympy==1.13.3
tifffile==2023.7.10
tqdm==4.65.2
typing_extensions==4.13.2
urllib3==2.2.3
zipp==3.20.2
